inventario-qr/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==2.7.4
python-dotenv==1.0.1
reportlab==4.2.0
qrcode==7.4.2
requests==2.31.0

# Linting/Formatting
ruff==0.4.4
black==24.4.2

# Testing
pytest==8.2.0
httpx==0.27.0 # For testing async client
